ui/requirements.txt
gradio>=3.0.0
requests
Pillow

